fixed MSVC compile warnings

+++ kernels/builders/heuristic_binning.h
@@ -337,7 +337,7 @@ namespace embree
 	float bestSAH = inf;
 	int   bestDim = -1;
 	int   bestPos = 0;
-	for (size_t dim=0; dim<3; dim++) 
+	for (int dim=0; dim<3; dim++) 
         {
           /* ignore zero sized dimensions */
           if (unlikely(mapping.invalid(dim)))
